fix: revert velocityproxy changes
